requirements.txt
django
djangorestframework==3.6.4
sqlalchemy
psycopg2
pyodbc
pyyaml
numpy
pandas
ipykernel
jsonschema
strict-rfc3339
# djangorestframework-csv
# djangorestframework-xml
# djangorestframework-yaml
# rfc3987